Add a "_cancel" target for buttons
When using javascript, the button detects if the history is not empty,
and then goes to the previous page visited. Otherwise it proceeds as
usual with the request object.

There is probably one issue: javascript's History object does not return
or throw exceptions when on the first page of the history and going
back. It just does nothing, and we cannot detect it (this is the reason
for the history length check in the first place).