Add preliminary support for CDN
With this change, it is possible to provide an absolute path when
delivering static content. This should help big sites with performance
(caching, latency...).

There is a design issue with this at the moment though: it should really
be the delivery engine (eg HTTPEngine) setting this value accordingly.
Likewise an engine could decide to provide a full local path (on the
filesystem) in order to locate resources for HTML files (when not
standalone, the current default).